feat: add ground truth columns to prediction output (#504)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->
Changes introduced:
- Using string enum type as opposed to int, so that the json output for
the enum field will be of the string value as opposed to int, which
would make it easier to understand the json content. This is important
for the use case of providing the enum values as input to flyte tasks.
- Instead of having all possible classes as separate attributes within
the inference schema class, they are now combined within a single field,
model_prediction_output. In addition, the prediction output is now
expected to implement the preprocess method, which will be used to
preprocess the dataframe before sending the result to Arize.
- Due to the existence of models that might have non-float prediction
value (such as ranking model), the subclasses of PredictionOutput are
expected to provide prediction types as well.
- Changes required for observation publisher and batch publisher, due to
the changes made to the Merlin SDK, will be on a separate pull request.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [x] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes