feat: Support model observability to pyfunc model (#502)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->
We introduce `PyFuncV3Model` as new pyfunc class that support model
observability. It is working fine for new model but it is not the case
for existing model that use `PyFuncModel` since merlin model can't be
changed once it is logged or deployed. Implication of this, users need
to create new model that use `pyfunc_v3` type and the endpoint will be
different hence there will be modification on the upstream caller. This
PR will allow `PyFuncModel` to user model observability as long as the
output of the model is `PyFuncOutput`.

This PR also allow `features`, `entities`, and `prediction` in
`PyFuncOutput` to be pandas DataFrame, previously the type is strictly
`Values` where in reality most of our users use pandas DataFrame and
numpy ndarray

Modification:
* api/api/validator.go -> Relax the validation to allow PyFuncModel also
* api/cluster/resource/templater.go -> Supply necessary env vars for
PyFunc and PyFuncV3 if model observability enabled
* python/sdk/merlin/pyfunc.py -> Supporting used of pandas DataFrame and
numpy ndarray for model input and output

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note

```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes