fix(api): Fix failing recovery of deployment attempts interrupted by API server pod termination (#517)
# Description
When a deployment attempt of a certain Merlin _model version_ (this
happens within an API server pod) gets interrupted midway by the
termination of the said API server pod, the attempt gets retried by one
of the other remaining API server pods or by a new one that gets spun up
to take its place. However this redeployment attempt (of the model
version) will currently fail:

![Screenshot 2024-01-11 at 6 00
26 PM](https://github.com/caraml-dev/merlin/assets/36802364/0665602f-6051-44be-b1df-4fce1b3fe42e)

Not only will the deployment attempt (shown as an entry in the
deployments table) show a status that will perpetually be stuck in
'pending', another deployment entry gets 'created' in the deployments
table that has the 'failed' state with an error message that claims that
the model version in question has already been deployed.

Trying to manually trigger a redeployment of the model version
subsequently would not work either and would instead only lead to the
same error message mentioned above to be returned. (The model version in
question has in reality been deployed but not known to the API server).

This PR thus introduces a fix to address the incorrect handling of
interrupted redeployments, so as 1) to stop the creation of an
additional deployment entry in the deployments table as well as resolve
occurrences of deployments stuck in the 'pending' state, and 2) to stop
the recovery/redeployment attempt from failing because an inference
service corresponding to the model version in question has already been
deployed.

# Context
The scenario described above can roughly be represented more accurately
by the following steps which take place in the Merlin API server (note
that _pod_ refers to the Merlin API server pod):
1. Deployment job gets picked up by pod 1 
2. Pod 1 creates a new entry in the deployments table
3. Pod 1 starts to deploy an inference service specified by the
deployment job
4. While waiting for the deployment to complete, pod 1 gets terminated
5. Deployment job (the same one) gets picked up by pod 2
6. **_Pod 2 creates an entry in the deployments table_**
7. **_Pod 2 starts to deploy an inference service specified by the
deployment job_**
8. The deployment fails because by this time the inference service
deployment started in step 3 is completed
9. The 2nd deployment entry in the deployments table is saved as failed
10. The deployment job gets saved as completed in the jobs table

The first fix to be introduced adds a check at step 6, to verify that
the last deployment entry in the DB for the model version is not stuck
in the ‘pending’ state (which would be unexpected). If so, we’ll assume
that the previous deployment attempt had failed and we would not create
a new entry in the DB. If there is no deployment entry found (the model
version has not been deployed before) or if the latest deployment entry
is not in the ‘pending’ state, we will create a new deployment entry
that corresponds to the new deployment attempt.

The next fix introduces an additional check in step 7, just before the
controller is used to create an inference service, to verify if there’s
already one that exists in the cluster where the model should be
deployed. If one already exists, we let the deployment workflow continue
as usual without creating a new one.

# Modifications
- `api/cluster/controller.go` - Add inference service checks before new
entities are created
- `api/queue/work/model_service_deployment.go` - Add deployment checks
before new entries get created in the DB
- `api/storage/deployment_storage.go` - Add new helper method
`GetLatestDeployment` which returns the latest Merlin model deployment
entry from the deployments table in the DB

# Tests
- Existing behaviour of redeployments should remain unchanged
- Added unit tests to the refactored controller, model service
deployment and deployment storage files

# Checklist
- [x] Added PR label
- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE
```