feat(api): Expose tolerations and node selectors for batch jobs (#539)
# Description
This PR introduces changes to expose tolerations and node selectors that
can be configured for batch jobs. Previously, these fields were
hardcoded and there was no way to change them other than by making
changes to the Merlin API server. With the new changes, these fields can
be configured via the Merlin API server config values file.

For this to be done, a new `batchJobTemplater` attribute of a new
`BatchJobTemplater` type is added to the batch controller, just like how
the real-time Merlin model controller
[has](https://github.com/caraml-dev/merlin/blob/683884464ced1da2565b9a1194bc370a17cb317b/api/cluster/controller.go#L94)
a `kfServingResourceTemplater` that propagates default configuration
values for real-time models to the model specs.

This new `BatchJobTemplater` struct only carries the new configs
corresponding to the tolerations and node selectors that are set on the
Merlin API server level, and 'adopts' the functions
`CreateSparkApplicationResource`, `createSpec`, `createDriverSpec` and
`createExecutorSpec` that were previously merely functions (i.e. these
are now made methods of the `BatchJobTemplater` struct).

🚨 This PR also adds a new GitHub action to e2e tests to clean up disk
storage on the GitHub runner to allow the e2e tests to pass.

# Modifications
- `api/batch/controller.go` - Addition of the `batchJobTemplater`
attribute to the batch controller
- `api/batch/resource.go` - Addition of the new `BatchJobTemplater`
struct and refactoring of its associated methods
- `api/config/batch.go` - Addition of a new `BatchConfig` struct to
contain the tolerations and node selectors to be added to batch jobs
- `.github/workflows/merlin.yml` - Adds new GitHub action to e2e tests
to clean up disk storage on the GitHub runner to allow the e2e tests to
pass

# Tests
- [x] Running new batch jobs with custom tolerations/node selectors

# Checklist
- [x] Added PR label
- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
None
```