fix: add Merlin project as config to observation publisher (#563)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->
# Description
<!-- Briefly describe the motivation for the change. Please include
illustrations where appropriate. -->
This PR addresses two bugs found in the observation publisher:
1. The model id config field in observation publisher is wrongly assumed
to be in the form of (project)-(model id), instead of just (model id) .
This affect the bigquery table write location for BigQuerySink, and
arize model id for ArizeSink.
2. Prediction log from standard model does not contain the "columns"
field, but this is not handled by the observation publisher.

# Modifications
<!-- Summarize the key code changes. -->
- API service will now supply model project as an input to the deployer.
- Observation publisher will have a new config field, project. This is
going to be passed to the observation sink as well.
- Observation publisher will now use the feature_orders field in
inference schema when trying to construct the pandas dataframe from the
prediction log, if the field value is non null

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [ ] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [ ] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
<!--
Does this PR introduce a user-facing change?
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note

```