chore(api): Update go dependencies in merlin api server (#551)
# Description
This MR basically bumps a couple of Go package dependencies used by the
Merlin API server, which in turn is imported by the Turing API server
and will have its own dependencies updated in a separate PR (the updates
in the PR will allow the dependencies of the Turing API server to also
be updated without having to be pinned to the versions used in the
Merlin API server). The version of Go used has also been bumped up from
1.20 to 1.22.

Main packages updated:
```
- github.com/GoogleCloudPlatform/spark-on-k8s-operator
- github.com/google/go-containerregistry 
- github.com/kserve/kserve
- k8s.io/api 
- k8s.io/apimachinery
- k8s.io/client-go
- knative.dev/networking
- knative.dev/pkg
- knative.dev/serving
- sigs.k8s.io/controller-runtime
```

Some minor refactoring was performed on structs that may have changed or
on functions called due to the changes in their expected argument -
these are annotated directly on the diffs.

# Modifications
- `api/cluster/controller.go` - Updated the KServe controller functions
as they now require a context
- `api/cluster/resource/templater.go` - Updated the `Handler` field of a
`Probe` struct to a `ProbeHandler` field
- `api/go.mod` - Updated various packages here

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
<!--
Does this PR introduce a user-facing change?
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```